+++ src/Admin/FieldDescriptionInterface.php
@@ -15,6 +15,11 @@ namespace Sonata\AdminBundle\Admin;
 
 /**
  * @author Thomas Rabaix <thomas.rabaix@sonata-project.org>
+ *
+ * @method string|null getTargetModel()
+ * @method bool        hasAdmin()
+ * @method bool        hasParent()
+ * @method bool        hasAssociationAdmin()
  */
 interface FieldDescriptionInterface
 {
@@ -106,8 +111,13 @@ interface FieldDescriptionInterface
 
     /**
      * Returns the parent Admin (only used in nested admin).
+     *
+     * @return AdminInterface|null // NEXT_MAJOR: Return AdminInterface
      */
     public function getParent(): ?AdminInterface;
+
+    // NEXT_MAJOR: Uncomment the following line
+    // public function hasParent(): bool;
 
     /**
      * Define the association mapping definition.
@@ -120,9 +130,15 @@ interface FieldDescriptionInterface
     public function getAssociationMapping(): array;
 
     /**
-     * Returns the related Target Entity.
+     * NEXT_MAJOR: Remove this method in favor of `getTargetModel()`.
+     *
+     * Returns the related Target object model.
+     *
+     * @deprecated since sonata-project/admin-bundle 3.69. Use `getTargetModel()` instead.
      */
     public function getTargetEntity(): ?string;
+
+    // public function getTargetModel(): ?string;
 
     /**
      * Sets the field mapping information.
@@ -157,8 +173,13 @@ interface FieldDescriptionInterface
 
     /**
      * Returns the associated Admin instance (only used if the field is linked to an Admin).
+     *
+     * @return AdminInterface|null // NEXT_MAJOR: Return AdminInterface
      */
     public function getAssociationAdmin(): ?AdminInterface;
+
+    // NEXT_MAJOR: Uncomment the following line
+    // public function hasAssociationAdmin(): bool;
 
     /**
      * Returns true if the FieldDescription is linked to an identifier field.
@@ -181,6 +202,9 @@ interface FieldDescriptionInterface
      * @return AdminInterface the admin class linked to this FieldDescription
      */
     public function getAdmin(): AdminInterface;
+
+    // NEXT_MAJOR: Uncomment the following line
+    // public function hasAdmin(): bool;
 
     /**
      * merge option values related to the provided option name.