+++ src/Datagrid/ListMapper.php
@@ -28,6 +28,10 @@ use Sonata\AdminBundle\Mapper\BaseMapper;
  */
 class ListMapper extends BaseMapper
 {
+    public const TYPE_ACTIONS = 'actions';
+    public const TYPE_BATCH = 'batch';
+    public const TYPE_SELECT = 'select';
+
     /**
      * @var FieldDescriptionCollection
      */
@@ -85,9 +89,9 @@ class ListMapper extends BaseMapper
             }
         }
 
-        // Ensure batch and action pseudo-fields are tagged as virtual
-        if (\in_array($type, ['actions', 'batch', 'select'], true)) {
-            $fieldDescriptionOptions['virtual_field'] = true;
+        // Type-guess the action field here because it is not a model property.
+        if ('_action' === $name && null === $type) {
+            $type = self::TYPE_ACTIONS;
         }
 
         if (\array_key_exists('identifier', $fieldDescriptionOptions) && !\is_bool($fieldDescriptionOptions['identifier'])) {
@@ -135,6 +139,11 @@ class ListMapper extends BaseMapper
         if (!isset($fieldDescriptionOptions['role']) || $this->admin->isGranted($fieldDescriptionOptions['role'])) {
             // add the field with the FormBuilder
             $this->builder->addField($this->list, $type, $fieldDescription, $this->admin);
+
+            // Ensure batch and action pseudo-fields are tagged as virtual
+            if (\in_array($fieldDescription->getType(), [self::TYPE_ACTIONS, self::TYPE_BATCH, self::TYPE_SELECT], true)) {
+                $fieldDescription->setOption('virtual_field', true);
+            }
         }
 
         return $this;